Upgraded Kafka, updated fixture setup for Azure Service Bus to clean namespace, eliminated some non-asserting tests

+++ tests/MassTransit.Azure.ServiceBus.Core.Tests/SubscriptionFilter_Specs.cs
@@ -67,7 +67,8 @@
 
             var managementClient = Configuration.GetManagementClient();
 
-            await managementClient.DeleteTopicAsync(topicName);
+            if (await managementClient.TopicExistsAsync(topicName))
+                await managementClient.DeleteTopicAsync(topicName);
 
             ServiceBusTokenProviderSettings settings = new TestAzureServiceBusAccountSettings();
